Add EEPROM programming capability. (#47)
This commit adds --eeprom command that works similarly to --flash. It performs
an EEPROM erase, writes contents of a hex file into EEPPROM and verifies that
the write was correct.

This command currently only works for 16k and 32k tinyAVR devices. For other
devices the correct eeprom_* constants in device/device.py need to be added.